refactor: move mocks to src folder
